#1 separando serviços para Alternative e Answer
