+++ src/Main.java
@@ -1,17 +1,18 @@
 import BE.domain.*;
-import BE.services.QuestionService;
-import BE.domain.Alternative;
+import BE.services.*;
 
 public class Main {
 
   public static void main(String[] args) {
     QuestionService questionService = new QuestionService();
+    AnswerService answerService = new AnswerService();
+    AlternativeService alternativeService = new AlternativeService();
 
     Question question = questionService.createQuestion(1,"1", "Quanto é 2 + 2", false);
-    Alternative alternative1 = questionService.createAlternative(1, question, "4", true, false);
-    Alternative alternative2 = questionService.createAlternative(1, question, "3", false, false);
-    Alternative alternative3 = questionService.createAlternative(1, question, "2", false, false);
-    Alternative alternative4 = questionService.createAlternative(1, question, "5", false, false);
+    Alternative alternative1 = alternativeService.createAlternative(1, question, "4", true, false);
+    Alternative alternative2 = alternativeService.createAlternative(1, question, "3", false, false);
+    Alternative alternative3 = alternativeService.createAlternative(1, question, "2", false, false);
+    Alternative alternative4 = alternativeService.createAlternative(1, question, "5", false, false);
 
 
     System.out.println(question.toString());
@@ -24,7 +25,7 @@ public class Main {
 
     //createAnswer
 
-    Answer answer = questionService.createAnswer(/*student,*/alternative1);
+    Answer answer = answerService.createAnswer(/*student,*/alternative1);
     System.out.println(answer);
   }
 }
